refactor(ruby): split constant/DSL machinery into constants.go
Move the bare-identifier call detection, constant collection/references,
and class-builder DSL detection to constants.go. Decompose collectConstRefs
(constRefSkipped + emitConstantRef + emitScopeResolutionRef) and collectConstants
(registerConstantAssignment + collectNestedConstants). The ruby package now
carries zero complexity-ledger entries. Behavior-preserving: 24 goldens
byte-identical.
